Separate web view vm tests and create a fixture for them
